Add files via upload
I calibrate an Vasicek model of Libor rates with coupon expiration 12-mont with data from 2000 to 2021
by algorithm from a paper on a calibrating the Vasicek model in attached material.  The main problem is that the function 
of mean value is not depending on t that is why estimation is not strong in case when rates has big volatility(then reversion coefficient 
attract process to a mean).  In future I am extremely interesting in implicating some stronger version of such model with mean value 
depending on a time which can be better calibration in that case.